refactor: separate auth, admin, and authenticated resource routes into individual files
