+++ backend/routes/auth.go
@@ -1,0 +1,22 @@
+package routes
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/kubestellar/ui/middleware"
+)
+
+// set up authentication routes
+func setupAuthRoutes(router *gin.Engine) {
+	// Authentication routes
+	router.POST("/login", LoginHandler)
+
+	// API group for all endpoints
+	api := router.Group("/api")
+
+	// Protected API endpoints requiring authentication
+	protected := api.Group("/")
+	protected.Use(middleware.AuthenticateMiddleware())
+	{
+		protected.GET("/me", CurrentUserHandler)
+	}
+}
